requirements.txt
argon2-cffi
argon2-cffi-bindings
asgiref
async-timeout
attrs
autobahn
Automat
billiard
boto3
botocore
celery
certifi
cffi
channels
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
constantly
cryptography
daphne
Deprecated
Django
django-ckeditor
django-cors-headers
django-environ
django-js-asset
django-redis
django-storages
djangorestframework
gunicorn
hyperlink
idna
incremental
jmespath
kombu
mutagen
packaging
Pillow
prompt-toolkit
psycopg2
pyasn1
pyasn1-modules
pycparser
pyOpenSSL
pyparsing
python-dateutil
pytz
redis
requests
s3transfer
service-identity
six
sqlparse
Twisted
txaio
typing_extensions
urllib3
vine
wcwidth
whitenoise
wrapt
zope.interface